[ref] renaming services into doc-api
