+++ src/app/commons/commons.module.ts
@@ -4,7 +4,7 @@ import { CommonModule } from '@angular/common';
 import { ToolComponent } from './tool/tool.component';
 import { DsPageItemModule } from './page';
 import { DsCodeTabsModule } from './code-tabs';
-import { DsDocApiModule } from './services/doc-api.module';
+import { DsDocApiModule } from './doc-api';
 import { DsStructureModule } from './structure';
 
 @NgModule({
